The coffeenoster theme & modified vi-mode for zsh
Coffeenoster takes most of it's inspiration from agnoster and the
agnoster theme for fish (oh-my-fish). The vi-mode-coffeevector is a
modified version of vi-mode, which adds a right-prompt. The difference
between the right-prompt in fish and vi-mode is that vi-mode includes
the full --INSERT-- and --NORMAL-- like how vim does, rather than the
shortened N and I. This addition was for clarity as well as aesthetics.